Refactor API and Bot Functionality with Enhanced Category and Record Management
- Updated `amount` column type in `records` table and models from `Integer` to `Float`.
- Added API endpoint to retrieve category by name with optional hidden flag.
- Refactored `APIClient` methods for user, category, and record management to improve type handling and error logging.
- Implemented new bot handlers for category management (`create`, `update`, `delete`) and expense recording, integrating with API.
- Replaced `bot_utils.py` with `helpers.py`, consolidating utility functions for inline keyboards and logging.
- Added new unit tests for category retrieval by name.
- Updated `compose.yaml` to use `postgres:16.2` and enable volume mounting for development.
- Introduced Redis connectivity checks and improved middleware for session and category management.

+++ alembic/versions/ffac1b7e140b_initial_migrations.py
@@ -57,7 +57,7 @@ def upgrade() -> None:
     )
     op.create_table('records',
     sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('amount', sa.Integer(), nullable=False),
+    sa.Column('amount', sa.Float(), nullable=False),
     sa.Column('user_id', sa.Integer(), nullable=False),
     sa.Column('category_id', sa.Integer(), nullable=False),
     sa.Column('currency', postgresql.ENUM('RUB', 'USD', 'EUR', name='currency', create_type=False), nullable=False),
